Use an uint32 instead of a long cast to evaluate the non-esp marker. This causes problems on 64bit architectures.
git-svn-id: svn://svn.shrew.net/ike/head@86 1bb156eb-19af-db11-b919-0014c23ffdf2

+++ source/iked/ike.io.network.cpp
@@ -218,7 +218,7 @@ long _IKED::loop_ike_nwork()
 						// check for NAT-T non-esp marker
 						//
 
-						unsigned long * marker = ( unsigned long * )( packet_udp.buff() + packet_udp.oset() );
+						uint32_t * marker = ( uint32_t * )( packet_udp.buff() + packet_udp.oset() );
 
 						if( !marker[ 0 ] )
 						{
